Refactor error toast messages
Resolves #1349
