+++ src/features/labels/Vote.tsx
@@ -4,7 +4,7 @@ import { arrowDownSharp, arrowUpSharp } from "ionicons/icons";
 import { voteOnPost } from "../post/postSlice";
 import React, { useContext } from "react";
 import { voteOnComment } from "../comment/commentSlice";
-import { downvotesDisabled, voteError } from "../../helpers/toastMessages";
+import { downvotesDisabled } from "../../helpers/toastMessages";
 import { PageContext } from "../auth/PageContext";
 import {
   calculateTotalScore,
@@ -18,6 +18,7 @@ import useHapticFeedback from "../../helpers/useHapticFeedback";
 import useAppToast from "../../helpers/useAppToast";
 import { formatNumber } from "../../helpers/number";
 import { styled } from "@linaria/react";
+import { getVoteErrorMessage } from "../../helpers/lemmyErrors";
 
 const Container = styled.div<{
   vote?: 1 | -1 | 0;
@@ -91,7 +92,11 @@ export default function Vote({
     try {
       await dispatch(dispatcherFn(id, vote));
     } catch (error) {
-      presentToast(voteError);
+      presentToast({
+        color: "danger",
+        message: getVoteErrorMessage(error),
+      });
+
       throw error;
     }
   }
